requirements.txt
asgiref==3.3.4
astroid==2.4.2
colorama==0.4.4
confusable-homoglyphs==3.2.0
da-vinci==0.3.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2
django-autoslug==1.9.8
django-autoslug-field==0.2.3
django-heroku==0.3.1
django-registration==3.1.1
django-thumbnails==0.4.0
django-widget-tweaks==1.4.8
gunicorn==20.1.0
insultgenerator==1.0.3
isort==5.7.0
lazy-object-proxy==1.4.3
loremipsum==1.0.5
mccabe==0.6.1
pi==0.1.2
Pillow==8.1.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2020.5
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
static-ranges==0.2.0
static3==0.7.0
toml==0.10.2
White-Noise==0.1.0
whitenoise==5.2.0
wrapt==1.12.1
